+++ backend/component/nlp/word_manage/base_dict.txt
@@ -1,5 +1,963 @@
+跳转
 蓝鲸
-王者荣耀
 英雄联盟
-pubg
-lol
+穿越火线
+自定义
+排查
+脚本
+报错
+找不到
+产品名称
+业务英文名
+英文缩写
+自由幻想
+地下城与勇士
+烽火战国
+幻想世界
+宠物企鹅
+音速
+宠物熊熊
+穿越火线
+炫舞
+丝路英雄
+七雄争霸
+七雄
+战地之王
+英雄联盟
+逆战
+御龙在天
+怪物猎人
+绿色征途
+洛克王国
+游戏人生
+英雄岛
+商业智能
+基础服务
+官网与营销
+研发部品质管理中心
+弹道轨迹
+王朝霸域
+互娱业务安全
+大乐斗2
+生化战场
+炫斗之王
+枪神纪
+刀剑2
+将魂
+怪物大作战
+战争前线
+天堂岛
+天堂1
+天堂2
+摩登城市
+九界
+三国塔防魏传
+节奏大师
+互娱开放平台
+万王之王3
+夜店之王
+剑灵
+疾风之刃
+炫舞时代
+天涯明月刀
+动漫平台
+数据安全组
+纵横九洲
+乐斗派
+花儿对对碰
+黄金矿工
+圣犬帕拉
+三国了没
+上古世纪
+自由足球
+时光小镇
+完美女孩
+腾讯游戏联盟
+动物大爆炸
+天天连萌
+天天爱消除
+天天酷跑
+疯狂联盟
+新终端开放平台
+软件包管理系统
+系统日志分析系统
+蓝鲸配置平台
+时间同步服务
+系统运维服务
+文件和版本服务
+备份归档系统
+防沉迷系统
+敏感词系统
+短信网关
+信息网关
+支付网关
+密保卡代理
+外包需求管理系统
+游戏下载服务
+互动娱乐研发部
+天天飞车
+天天炫斗
+新终端游戏官网平台
+互娱效能协同中心
+互娱经营分析系统
+农牧场
+神魔之塔
+全民切水果
+天天星连萌
+刀锋铁骑
+腾讯文学
+斗战诛天
+游戏圈监测系统
+全民英雄
+王牌对决
+神之浩劫
+全民泡泡大战
+进击的女神
+宝宝斗场
+全民水浒
+全民烽火
+欢乐西游
+全民砰砰砰
+天天富翁
+互娱游戏论坛
+全民飞机大战
+傲世西游
+铁将军
+蓝鲸
+互娱监控平台
+天天德州
+城堡争霸
+雷霆战机
+战争总动员
+市场部前端重构组
+全民精灵
+全民小镇
+全民炫舞
+手机游戏门户
+全民打怪兽
+全民偶像-中国好声音
+糖果传奇
+天天宝石大战
+全民农场
+全民斗三国
+市场部公共开发组
+天天来塔防
+天天风之旅
+九龙战
+海纳中心业务
+变形金刚
+七骑士
+手游诊断系统
+天天斗三国
+超神英雄
+腾讯游戏帮帮
+英雄传奇
+互娱数据传输系统
+无尽之剑-命运
+道聚城
+仙剑奇侠传官方手游
+复仇者联盟
+英雄杀
+终端欢乐斗牛
+终端欢乐麻将
+终端欢乐斗地主
+猛击三国
+游龙英雄
+仙侠道
+微岛
+全民突击
+微瓴
+勇者大冒险
+天天来战
+天天传奇
+全民破坏神
+脑力达人
+魔龙与勇士
+兽人必须死
+全民闯天下
+宝贝童话
+大闹三国
+洛克精灵战记
+宫爆老奶奶2
+手游英雄杀
+天天守卫战
+腾讯游戏众测平台
+驯龙日记
+潮爆三国
+欢乐西游外发版
+全民无双
+全民水浒外发版
+游戏手机助手
+豆豆总动员
+战斗吧剑灵
+独立防线
+部落守卫战手机版韩版
+新终端集群服务
+独立防线北美版
+天魔幻想
+语音告警屏蔽调试
+蜀山传奇
+神枪手
+街机传奇
+待删除_六界仙尊
+烈火屠龙
+待删除_斩仙
+焚天
+苍穹变
+英魂之刃
+暴风王座
+快意江湖杀
+真魔
+传奇时代
+龙骑士传
+称霸天下
+枪林弹雨
+待删除_迷你帝国
+小师妹
+待删除_横扫西游
+神将2
+格斗大冒险
+笑傲江湖
+业务受理
+业受
+全民打坦克
+大皇帝
+手机网络质量保障平台
+火影忍者手游
+轩辕世界
+仙魂
+天珠
+疾风之刃手机版
+待删除_七雄争霸手游
+待删除_三国之刃
+蓝鲸数据平台对外版
+王牌战机
+智龙迷城
+奇迹暖暖
+全民主公
+全民枪王
+全民超神
+拳皇98
+天天飞车北美版
+三国笑传
+好客系统
+魔方质量管控组
+腾讯智能玩具
+荣耀之战
+热血传奇
+蓝鲸数据平台对内版
+命运之旅
+魂斗罗
+迷你帝国
+七雄争霸手游
+横扫西游
+三国之刃
+全民斩仙
+凤舞江湖
+古剑奇谭
+灵山奇缘
+龙破九天之龙啸天下
+保卫萝卜3_新赛季
+传奇霸业
+全民斗场
+星河战神
+无双剑姬
+水晶大冒险
+新流星搜剑录
+轮回
+质量开放平台
+王者荣耀
+天天象棋
+欢乐升级
+全民斗战神
+腾讯桌球
+捕鱼来了
+书剑恩仇录
+时空猎人
+寻龙诀之黄金帝国
+镇魔
+魔域破坏神
+剑曲
+客厅早教游戏
+极限三国
+游戏质量平台
+战争霸业
+九星诀
+腾讯客户端软件质量监控平台
+王牌传说
+召唤师对决
+魔剑之刃
+三国战神
+巨炮战队
+霸天之刃
+我欲封天
+赤壁战神
+帝王霸业
+逍遥
+萌仙页游
+血战魔龙
+部落大乱斗
+天天炫斗越南版
+天天炫斗英文版
+天天炫斗韩国版
+天天幻灵
+三国乱世
+扩缩容资源池
+梦想仙侠
+微宝
+台球大师
+公众号运营平台
+穿越火线_手游
+天书残卷
+全民突击-韩国版
+潘多拉
+新御剑天涯
+魔尊
+无上神兵
+口袋猛汉
+三国战纪
+真无双乱舞
+天天桥牌
+欢乐五子棋
+影视平台
+御龙在天手游
+真天龙八部
+全民卡牌大师
+神犬小七
+怪物猎人开拓记
+三国群英战
+龙城铁骑
+故障自愈
+花千骨
+龙珠激斗
+傲世群侠
+苍穹变手游
+怪物城堡
+天堂2-点击屠龙
+欢乐拼三张
+斗破苍穹2015
+英雄战歌
+琅琊榜
+火影忍者_台湾版
+天天军棋
+蓝鲸作业平台
+蓝鲸管控平台
+海外手游公共系统
+佛本是道
+神魔怒
+疯狂点击
+泡泡堂手机版
+火源计划
+仙侠道2
+街机炮炮堂
+云中歌
+全民主公页游版
+坦克大战
+泰坦世界
+蓝鲸作业平台内部版
+蓝鲸公共组件
+星际火线
+热血江湖传
+腾讯围棋
+风暴大陆
+独立防线外发版
+开放的传统游戏数字图书馆
+火线英雄
+勇者盟约
+铁骑神兵
+三国编年史
+诸神黄昏
+灵域
+腾讯影业泛娱乐平台
+大战神
+[外发]火影忍者台湾版
+[外发]独立防线北美版
+外发九龙战台湾版
+[外发]官网与营销
+盟军防线
+腾讯玩物圈
+冒险岛2
+女神联盟2
+神印王座
+火影忍者_德国版
+偶像梦幻祭
+剑侠情缘
+御龙在天经典版
+小鱼飞飞
+天天传奇日本版
+腾讯游戏开发者已废弃
+游戏音频支持
+腾讯云蓝鲸
+征途
+流放之路
+圣斗士星矢卡牌
+光子玩家体验服务
+掌上游戏社区
+腾讯游戏加速平台
+轩辕传奇手游
+微扑克
+战国之怒
+武神赵子龙
+剑雨江湖
+全民突击欧美版本
+欢乐坦克
+龙主霸天
+外发九龙战韩国版
+传奇世界手游
+天龙八部手游
+剑网3指尖江湖
+傲战无双
+传奇盛世
+第一舰队
+九阴绝学
+电视大厅业务
+独立防线韩国版
+皇图
+[外发]火影忍者_德国版
+心悦俱乐部
+天天传奇新马版
+弹弹堂
+红警之坦克风暴
+全民突击-日本版
+全民超神繁体中文版
+数据挖掘
+梦幻诛仙
+全民斗战神外发版
+麻将连连看
+魔龙之戒
+守卫雅典娜
+主宰西游
+腾讯游戏学院
+欢乐飞行棋
+四大名捕
+糖果泡泡乐
+神座
+幻兽起源
+神州霸业
+猎鱼达人
+下载管理平台
+火影忍者_葡语版
+天火大道
+游戏体验管理平台
+明星三缺一
+通天西游
+轩辕剑正版
+冷兔跑酷
+足球经理2
+龙之谷
+花样击球
+神鬼无双
+西游破坏神
+天堂2-誓言
+欢乐拼点
+王牌斗地主
+剑灵页游
+腾讯游戏助手
+手游网络优化
+掌上英雄联盟
+泡泡达人
+天天传奇泰语版
+战姬无双
+海外天天德州东南亚版
+海外天天风之旅东南亚版
+海外全民飞机大战国际版
+海外天天爱消除国际版
+海外天天飞车东南亚版
+皇室战争
+乱世王者
+武极天下正版
+决战武林
+绝对三国
+英雄无敌
+宝贝听听
+剑侠情缘兵器谱
+全民突击台湾版
+全民突击泰国版
+全民突击新马版
+互娱资源公共平台
+资源运营服务
+疯狂贪吃蛇
+寻仙手游
+开心泡泡猫
+九州天空城
+唐门六道
+不朽传奇
+火影忍者-美东
+火影忍者-美西
+火影忍者-欧洲版
+火影忍者-东南亚版
+消除者联盟
+天涯明月刀韩国版
+天空追逐者
+神奇星历险记
+卡罗大冒险
+永无止境
+龙之灵
+梦幻恋舞
+完美漂移
+火箭联盟
+魔法王座
+青云志
+幻城
+封神英雄榜
+剑倚天下
+上古伏魔录
+千军纵横
+秀丽江山长歌行
+斗罗大陆
+大唐盛世
+盗墓笔记
+奇迹重生
+华夏手游版
+新六界仙尊
+雷霆战纪
+欢乐球吃球
+热血战歌
+天庭战争
+公共技术研发中心
+天涯明月刀手游
+海王宝藏
+梦想星城
+街篮高手
+火影忍者_法语版
+魔魔打勇士
+热血传奇网页版
+暴击英雄
+水族馆消消乐
+命运方舟
+蓝月传奇
+魔塔奇兵
+萌将
+蛇蛇争霸
+全民弹弹堂
+天龙八部页游版
+狂热战车
+一起来捉妖
+街头篮球
+游戏美术资源商城
+王牌捉妖师
+石器时代手游
+单机游戏通用服务
+穿越火线手游海外版
+蛋蛋总动员
+守望先锋
+塔防三国志
+传奇霸业手游版
+你画我猜
+拳皇命运
+国王指意
+饥荒联机版
+互娱备份系统平台
+御龙在天美人版
+海底大作战
+玫瑰骑士团
+猪来了
+穿越火线微端
+夏目的美丽日记
+通道晋升机考专用
+全民大乐斗
+从前有座灵剑山之格斗修仙传
+魔力宝贝
+我的王朝
+八爪鱼
+捕鱼大玩咖
+全民大航海
+大圣快点
+部落冲突
+无尽争霸
+小米超神
+时空召唤
+虚荣
+帕拉贡
+公主战争
+智营解析
+御史大冒险
+幸运捕鱼
+阴阳师
+大话西游
+梦幻西游
+问道
+神武2
+麻将来了
+战地联盟
+重返战场
+僵尸大作战
+穿越火线手游泰国版
+穿越火线手游越南版
+互娱数据中心
+游戏持续集成体验
+大青云
+少年群侠传
+天域苍穹1
+火爆拼三张
+星球大乱斗
+完美红颜
+炫舞梦工厂
+全民岛主
+兽人必须死手游
+捕鱼娱乐城
+全民打雪球
+择天记
+弹弹来了
+火影忍者_西语_拉美版
+火影忍者_西语_欧洲版
+奇迹觉醒
+迷你世界
+无尽远征
+迷你血战2电视版
+腾讯游戏成长守护平台
+三国群英传-霸王之业
+跑跑斗地主
+腾讯云
+腾讯微享-待删除
+腾讯微享
+妖尾
+太极熊猫
+功夫熊猫
+初音未来-梦幻歌姬
+掌上穿越火线
+特战英雄
+统一支付流水
+天龙八部手游体验服
+自由幻想手游
+极光游戏
+腾讯出海宝
+支撑产品
+云裳羽衣
+天天斗地主
+光子技术中心
+卡片怪兽
+舰炮与鱼雷
+凯蒂环球之旅
+游戏公众号运营平台
+全民大富豪
+反恐精英_全球攻势
+大掌门2
+火力对决
+疯狂动物城-筑梦日记
+火影忍者-波兰版
+狐妖小红娘
+倩女幽魂
+企鹅电竞
+贪吃小怪物
+剑侠情缘2
+纪念碑谷2
+火影忍者-韩语版
+御龙在天手游港澳台
+腾讯国际象棋
+饭局狼人杀
+王者荣耀电视台
+魂武者
+文化产业
+穿越火线手游台湾版本
+金枪鱼
+枪神传说
+进化娱乐工作室
+妖神记
+灯塔业务
+死神来了
+使命召唤手游海外版
+欢乐三张牌
+蓝鲸运营
+谍战项目
+腾讯广东麻将
+腾讯贵州麻将
+跑跑卡丁车
+全民突击越南版
+腾讯四川麻将
+蓝鲸监控
+冒险岛正版手游
+绝地求生
+火影忍者-俄罗斯版
+超神乱斗
+腾讯斗地主
+蓝鲸爬虫
+幸运之夜
+腾讯湖南麻将
+暗黑破坏神3
+火影忍者-土耳其版
+蓝鲸集成
+光荣使命
+欢乐消消消
+魔弦传说
+火影忍者_泰国
+全民小镇网页版
+外部游戏助手
+黑黑
+弹弹堂大冒险
+保卫萝卜迅玩版
+轻松跑得快
+大家来找茬
+欢乐坦克大战
+腾讯双扣
+魂斗罗体验服
+乱世王者亚太版
+反恐行动
+战舰世界
+枪战英雄
+生死狙击
+火线精英
+坦克世界
+奥普赛
+愤怒的小鸟梦幻岛
+罗博造造
+秦时明月手游
+蓝盾
+一起来飞车
+无限法则
+松松总动员
+狂野飚车
+速度与激情8
+跑跑卡丁车手机版竞品
+狂野飙车8.极速凌云
+极品飞车.无限狂飙
+
+一人之下
+御天传奇
+凯蒂环球之旅外发版
+绝地求生-全军出击
+终结者2_审判日
+小米枪战
+荒野行动
+剑网三
+逆水寒
+新倩女幽魂
+绝地求生-刺激战场
+大天使之剑
+初音速
+决战平安京
+奇迹最强者
+寻仙2
+非人学院
+超维对决
+天天爱画猜
+女皇陛下
+侍魂-胧月传说
+腾讯游戏海外加速平台
+疯狂贪吃蛇国际版
+楚留香手游
+一刀流
+贪玩蓝月微端
+决战沙城
+传奇战域
+烈焰手游
+烈焰龙城
+倩女幽魂手游
+烈焰修罗
+烈焰之刃
+烈焰皇朝
+怒火一刀
+嗜血攻沙
+屠龙战
+屠龙战神
+王者传奇
+我们的沙城
+电击文库-零境交错
+代号_龙
+堡垒之夜
+自由禁区
+光荣使命体验服
+仙剑奇侠传四
+斗破苍穹正版手游
+天天王者
+拉结尔
+云梦四时歌
+梦幻家园
+梦幻花园
+全职猎人
+捉妖联盟
+腾讯游戏精品加速平台
+穿越火线手游北美版本
+穿越火线手游西欧版本
+恋与制作人
+荒野乱斗
+轩辕东南亚国际版
+轩辕港澳台繁体版
+足球经理移动版2018
+家长守护
+完美世界
+决战-蜜獾
+轩辕马来西亚简体版
+轩辕泰国泰文版
+我的王朝外发版
+恋爱48天
+弹弹魔法国
+光明勇士
+魂斗罗东南亚版本
+用户研究与体验设计中心-部门级
+欧氏几何
+挑战计算器
+蜡烛人
+全民冠军足球2018
+全军出击国际版
+斗地主来了
+穿越火线高清竞技大区
+欢乐五子棋小游戏
+节奏大师小游戏
+万智牌-竞技场
+王者围城
+王朝霸域小游戏
+执剑之刻
+游戏平台部数据中心
+浪漫玫瑰园
+烽火战国小游戏
+猜猜谁说谎
+夜店之王小游戏
+阿凡达
+欢乐飞行棋--印度版
+中台通用平台组件
+五子大作战
+全军出击国际版体验服
+育碧小游戏
+电竞直播
+野生之地
+刺激战场体验服
+麻将来了小程序
+3366游戏
+远离豆豆
+元气美食战
+飙酷车神2
+穿越火线手游印度版本
+穿越火线手游印尼版本
+穿越火线手游巴西版本
+穿越火线手游俄罗斯版本
+小飞机大战
+权力的游戏
+流言侦探-活着的死者
+赛尔号腾讯版
+银河掠夺者
+使命召唤手游国内版
+爆裂碎块
+蓝鲸应用开发组
+弹珠地牢
+功能游戏
+多米诺
+我的起源
+全民小镇小游戏
+英雄杀小游戏
+环形防御
+欢乐捕鱼大战
+王牌战士
+疯狂打砖块
+腾讯欢乐反斗星
+王者荣耀体验服
+疯狂击球手
+摇摆鲸鱼
+海水砰砰砰
+堆高高
+闹闹天宫
+激战狂潮
+恋之农场小程序
+漫威超级争霸战
+魂斗罗韩国版本
+篝火营地
+莺知夏浅
+手工星球
+文明大爆炸
+大家来找茬小游戏
+果冻派对
+互娱讲师认证
+龙珠最强之战
+天涯明月刀港澳台版本
+飞机跳跳跳
+武士卡尔
+豆豆蛇
+生存日记
+全民突击小游戏
+海滨消消乐小游戏
+跃动篮球
+腾讯游戏小悦
+欢乐猜拳大战
+弹个球
+金属对决国内
+仙境传说-爱如初见
+肿瘤医师
+小森生活
+时空射手
+少女咖啡枪2
+猪来了之环球旅行
+英雄就是我
+代号夏娃
+全民爱公益
+食物语
+风暴魔域
+狐妖小游戏
+移动大厅自研小游戏
+子曰诗云
+萌宠消方块
+脑力达人小游戏
+互动故事
+全民主公2
+牌师
+电竞工坊
+众生
+存储计算平台
+猫咪别追我
+螺旋风暴
+欢乐割绳子
+中国式家长
+银之守墓人-对决
+海盗法则
+逍遥诀
+彩虹六号_围攻
+临行之日
+脸黑先生
+洛菲斯的呼唤
+进化之地2
+腾讯即玩
+腾讯听听智能音箱
+欢乐坦克大战2
+熊猫博士玩具车
+熊猫博士亚洲餐厅
+熊猫博士巴士司机
+熊猫博士的冰淇淋车
+真龙霸业日本
+真龙霸业港澳台东南亚韩国
+我的英雄学院_最强英雄
+点击暴富小程序
+如懿传
+文明火种
+腾讯双扣小游戏
+节奏丛林
+灵魂之火
+荣耀战魂
+汤姆克兰西-全境封锁2
+折扇
+幽林怪谈
+影视业务
+社区业务
+创意设计中心业务
+用研业务
+公共业务
+小小宫匠
+纳木
+乐部落
+北冕
+奇妙世界
+品管实验室
+圣斗士星矢海外
+螺旋风暴_海外
+使命召唤-黑色行动4
+尖塔奇兵
+星露谷物语
+英雄战歌小游戏
+北方绝境
+圣斗士星矢香港
+圣斗士星矢台湾
+使命召唤-黑色行动4-香港
+欢乐小游戏合集
+合金弹头
+未转变者
+轩辕小游戏
+蓝鲸管控平台编译服务
+欧呜欧小游戏
+嘉期
+电是怎么形成的
+星辰平台
+品质管理平台
+第三方
+用户名
+跳回
+不能
+离线
+流量
