+++ backend/intent/plugins/devops/api.py
@@ -13,15 +13,11 @@ either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
 
-WXBIZMSGCRYPT_OK = 0
-WXBIZMSGCRYPT_VALIDATESIGNATURE_ERROR = -40001
-WXBIZMSGCRYPT_PARSEXML_ERROR = -40002
-WXBIZMSGCRYPT_COMPUTESIGNATURE_ERROR = -40003
-WXBIZMSGCRYPT_ILLEGALAESKEY = -40004
-WXBIZMSGCRYPT_VALIDATECORPID_ERROR = -40005
-WXBIZMSGCRYPT_ENCRYPTAES_ERROR = -40006
-WXBIZMSGCRYPT_DECRYPTAES_ERROR = -40007
-WXBIZMSGCRYPT_ILLEGALBUFFER = -40008
-WXBIZMSGCRYPT_ENCODEBASE64_ERROR = -40009
-WXBIZMSGCRYPT_DECODEBASE64_ERROR = -40010
-WXBIZMSGCRYPT_GENRETURNXML_ERROR = -40011
+from opsbot import CommandSession
+
+
+async def parse_user_id(session: CommandSession):
+    r = await session.bot.call_action('tencent/user/convert_to_name',
+                                      userid_list=[session.ctx['msg_sender_id']])
+
+    return r.get('user_list')[0]['name'] if r.get('user_list') else session.ctx['msg_sender_id']
