+++ backend/opsbot/log.py
@@ -11,9 +11,7 @@ software distributed under the License is distributed on
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
 either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
-"""
 
-"""
 Provide logger object.
 
 Any other modules in "opsbot" should use "logger" from this module
