+++ backend/opsbot/plugins/base.py
@@ -13,16 +13,10 @@ either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
 
-from opsbot import on_command, CommandSession, permission as perm
-from opsbot.message import unescape
+from opsbot import on_command, CommandSession
 
 
 @on_command('echo')
 async def echo(session: CommandSession):
     await session.send(session.state.get('message') or session.current_arg)
 
-
-@on_command('say', permission=perm.SUPERUSER)
-async def say(session: CommandSession):
-    await session.send(
-        unescape(session.state.get('message') or session.current_arg))
