+++ backend/protocol/xwork/exception.py
@@ -12,14 +12,3 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
 either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-
-"""
-Xwork configurations.
-"""
-
-CORPID = ""
-FWID = ""
-SERVICE_ID = ""
-SECRET = ""
-TOKEN = ""
-AES_KEY = ""
