+++ backend/requirements.txt
@@ -35,6 +35,7 @@ urllib3==1.26.4
 wsproto==1.0.0
 yarl==1.6.3
 redis==3.5.3
+xmltodict==0.12.0
 python-Levenshtein==0.12.2
 gensim==4.0.1
 jieba==0.42.1
