+++ docs/release.md
@@ -1,7 +1,10 @@
 # 版本日志
 
-## 1.0.0
+## 1.1.0
 
 - 新功能
-  - 开放基础功能
+  - 支持多协议适配器
+  - nlp模块解析改为异步执行
+  - 支持蓝盾流水线
+
